requirements.txt
streamlit==1.40.2
python-dotenv==1.0.1
pdfplumber==0.11.4
psutil==6.0.0
langchain==0.3.9
langchain-anthropic==0.3.0
langchain-openai==0.2.11
langchain-community==0.3.9
langchain-core==0.3.21
langchain-experimental==0.3.3
langgraph==0.2.56
langsmith>=0.1.17,<0.2.0
faiss-cpu==1.9.0
sentence-transformers==3.2.1
rank-bm25==0.2.2
boto3==1.35.74
pydantic==2.10.3
typing-extensions==4.12.2
anthropic>=0.39.0,<1.0.0
openai>=1.54.0,<2.0.0